+++ packages/integrations/mdx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @astrojs/mdx
 
+## 5.0.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @astrojs/markdown-remark@7.0.0-beta.10
+
 ## 5.0.0-beta.10
 
 ### Patch Changes