+++ packages/integrations/netlify/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @astrojs/netlify
 
+## 7.0.0-beta.13
+
+### Patch Changes
+
+- Updated dependencies [[`745e632`](https://github.com/withastro/astro/commit/745e632fc590e41a5701509e9cc4ed971bdddf74)]:
+  - @astrojs/internal-helpers@0.8.0-beta.2
+  - @astrojs/underscore-redirects@1.0.0
+
 ## 7.0.0-beta.12
 
 ### Minor Changes